building.js: working on add and get buildings

+++ src/main/java/com/devmountaincapstone/propertyportal/dtos/BuildingDto.java
@@ -14,7 +14,7 @@ import java.util.Set;
 @NoArgsConstructor
 public class BuildingDto implements Serializable {
     private Long id;
-    private Integer buildingNumber;
+    private Long buildingNumber;
     private PropertyDto propertyDto;
     private Set<UnitDto> unitDtoSet = new HashSet<>();
